add router and edition container
move logic for rendering an edition to a new edition container

closes #7
