--- src/App.js
@@ -1,29 +1,0 @@
-import React from 'react';
-import Header from './components/Header';
-import './App.css';
-
-const App = () => {
-
-  const publishDate = "12/12/2016"; // TODO: use dynamic date
-  const onApprove = () => alert("approve"); // TODO: Toggle approved state
-  const onInfo = () => alert("info"); // TODO: Toggle info menu
-  const onNext = () => alert("next"); // TODO: Navigate to next date
-  const onPrevious = () => alert("previous"); // TODO: Navigate to prev date
-
-  return (
-    <div className="App">
-      <div className="container">
-        <Header
-          isApproved={false}
-          onApprove={onApprove}
-          onInfo={onInfo}
-          onNext={onNext}
-          onPrevious={onPrevious}
-          publishDate={publishDate} />
-      </div>
-    </div>
-  );
-
-}
-
-export default App;
