+++ src/App.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import Router from 'react-router';
+import './App.css';
+
+const App = ({children}) => {
+
+  return (
+    <div className="App">
+      <div className="container">
+        {children}
+      </div>
+    </div>
+  );
+
+}
+
+export default App;
