+++ src/index.js
@@ -1,9 +1,17 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
+import { Router, Route, hashHistory } from 'react-router'
 import App from './App';
+import Editions from './containers/Editions';
+import Edition from './containers/Edition';
 import './index.css';
 
 ReactDOM.render(
-  <App />,
+  <Router history={hashHistory}>
+    <Route path="/" component={App}>
+      <Route path="/editions" component={Editions} />
+      <Route path="/editions/:publishDate" component={Edition} />
+    </Route>
+  </Router>,
   document.getElementById('root')
 );
